refactor(engine): code review: drop `ScheduledTrigger` and restore `JobTimeoutTrigger`
