+++ engine/src/main/java/io/camunda/zeebe/engine/processing/job/JobTimeoutTrigger.java
@@ -10,35 +10,65 @@ package io.camunda.zeebe.engine.processing.job;
 import static io.camunda.zeebe.util.sched.clock.ActorClock.currentTimeMillis;
 
 import io.camunda.zeebe.engine.processing.streamprocessor.ReadonlyProcessingContext;
+import io.camunda.zeebe.engine.processing.streamprocessor.StreamProcessorLifecycleAware;
 import io.camunda.zeebe.engine.processing.streamprocessor.writers.TypedCommandWriter;
 import io.camunda.zeebe.engine.state.immutable.JobState;
 import io.camunda.zeebe.protocol.record.intent.JobIntent;
+import io.camunda.zeebe.util.sched.ScheduledTimer;
 import java.time.Duration;
 
-public final class JobTimeoutTrigger extends ScheduledTrigger {
+public final class JobTimeoutTrigger implements StreamProcessorLifecycleAware {
   public static final Duration TIME_OUT_POLLING_INTERVAL = Duration.ofSeconds(30);
   private final JobState state;
 
+  private ScheduledTimer timer;
   private TypedCommandWriter writer;
+  private ReadonlyProcessingContext processingContext;
 
   public JobTimeoutTrigger(final JobState state) {
     this.state = state;
   }
 
   @Override
-  protected Duration getPollingInterval() {
-    return TIME_OUT_POLLING_INTERVAL;
-  }
-
-  @Override
-  protected void executeOnTrigger() {
-    deactivateTimedOutJobs();
-  }
-
-  @Override
   public void onRecovered(final ReadonlyProcessingContext processingContext) {
-    super.onRecovered(processingContext);
+    this.processingContext = processingContext;
+    timer =
+        this.processingContext
+            .getActor()
+            .runAtFixedRate(TIME_OUT_POLLING_INTERVAL, this::deactivateTimedOutJobs);
     writer = processingContext.getLogStreamWriter();
+  }
+
+  @Override
+  public void onClose() {
+    cancelTimer();
+  }
+
+  @Override
+  public void onFailed() {
+    cancelTimer();
+  }
+
+  @Override
+  public void onPaused() {
+    cancelTimer();
+  }
+
+  @Override
+  public void onResumed() {
+    if (timer == null) {
+      timer =
+          processingContext
+              .getActor()
+              .runAtFixedRate(TIME_OUT_POLLING_INTERVAL, this::deactivateTimedOutJobs);
+    }
+  }
+
+  private void cancelTimer() {
+    if (timer != null) {
+      timer.cancel();
+      timer = null;
+    }
   }
 
   void deactivateTimedOutJobs() {
